Refactor
Move Option1Parser and Option2Parser into their own files
Move DB into its own file
Create util.py for helping functions
