+++ parser.py
@@ -1,18 +1,7 @@
-import os
 import sys
 
 from model import Todo
-
-
-def clear_screen():
-    _ = os.system('cls' if os.name == 'nt' else 'clear')
-
-
-def print_activity(activity):
-    id, task, created_at, updated_at = activity
-
-    row = f"{id} - {task} - ({created_at} - {updated_at if updated_at else 'not updated'})"
-    print(row)
+from util import clear_screen, print_activity
 
 
 class CoreParser:
@@ -112,113 +101,3 @@ class CoreParser:
         else:
             clear_screen()
             sys.exit()
-
-
-class Option1Parser:
-    '''
-    1 - add an activity
-    2 - read an activity
-    3 - read all activities
-    4 - update an activity
-    5 - delete an activity
-    6 - delete all activities
-    7 - exit
-    clear - clear the screen
-    doc - display simple tutorial
-    '''
-
-    def __init__(self):
-        self.requires_input = True
-        self.print_message()
-        self.input_activity()
-
-    def input_activity(self):
-        self.command = input(">>> Enter command option: ")
-        self.id = 0
-        self.task = ""
-
-        if self.command in ['2', '4', '5']:
-            self.id = input(">>> Enter task id: ")
-
-        if self.command in ['1', '4']:
-            self.task = input(">>> Enter task: ")
-
-    def print_message(self):
-        message = '''
-            1 - add an activity
-            2 - read an activity
-            3 - read all activities
-            4 - update an activity
-            5 - delete an activity
-            6 - delete all activities
-            7 - exit
-            clear - clear the screen
-            doc - display simple tutorial
-            '''
-        print(message)
-
-
-class Option2Parser:
-    '''
-    - add activity
-    - read activity_id
-    - read *
-    - read all
-    - update activity_id new_task
-    - delete activity_id
-    - delete *
-    - delete all
-    - clear - clear the screen
-    - doc - display simple tutorial
-    '''
-
-    def __init__(self, parsed_input):
-        self.requires_input = False
-        self.command = ""
-        self.id = 0
-        self.task = ""
-
-        self.input_activity(parsed_input)
-
-    def print_message(self):
-        tutorial = '''
-        This is a simple tutorial on how to run app on terminal
-
-        add "Call John Doe at 2pm"
-        (This is adds a new activity)
-
-        read 2
-        (This reads an activity with id=2)
-
-        read all
-        read *
-        (This reads all the activities - the id is all or *)
-
-        update 2 "Call John Doe at 3pm"
-        (This update an activity with id=2 with the text as task)
-
-        delete 2
-        (This deletes an activity with id=2)
-
-        delete all
-        delete *
-        (This deletes all there activities - the id is all or *)
-
-        clear - clears the screen
-
-        doc - display simple tutorial
-        '''
-        print(tutorial)
-
-    def input_activity(self, input_data):
-        
-        self.command = input_data[0]
-
-        if self.command == 'add':
-            self.task = input_data[1]
-
-        if self.command in ['read', 'update', 'delete']:
-            self.id = input_data[1]
-
-        if self.command == 'update':
-            self.task = input_data[2]
